Reinstate drive to robot.
Looking back at last year's code, it was a mess. Packages were created
that ulimately lead to nothing, the code is trying to access files that
do not exist, etc. While there is some things that I feel would be good
to return to and work on, I think it would be a better idea to rebuild
the code.

This should be a quick rebuild, taking things that we wanted to do and
adding them to this new version. Along with this, starting fresh will
allow for easier development in the future.

This code should make it so that a person can drive the robot. Note
that I did not test it out (finished it at home)- might need to be
changed.