Add ball feeder functionality to the robot
During the last weeks before break, the team (or at least the members
that were still working on the robot) decided that the robot needed
a "feeding mechanism" for the balls that would be shot out of the
robot in order to prevent jamming from occuring when loading the
balls in. This "feeding mechanism" is a small wheel above the
loading chute that will feed balls toward the shooter at the
end of said chute.

In order to do this, I added a new command and subsystem for the
ball feeder (since the command has to be run alongside the
command for the shooter) as well as a new button to run this
new command.