Fix differential drive update issues during teleop
When originally testing out the code for the robot, WPILib
would hang up due to a DifferentialDrive error. This was due to the
fact that the differential drive for the robot was not being updated
enough. In order to fix this, I changed the code so that the
differential drive would be updated using a joystick (in contrast
to a Xbox controller).