FRC 2020 Season Code Branch
This is the code from the 2020 FRC Season. It is unfinished and has some
issues functionality wise. The major issue is that some parts of the
robot may not line up with the code (whether that be that we are missing
something or the robot is missing something which was implemented here).

The 2019-2020 FRC Season was incredibly hectic for the team, not only
because of the pandemic, but due to our ambitious design for the robot
that year. The timing of the pandemic could not have been worse for
software especially, because it hit just as we began to go into full
throttle to finish the code. Hopefully this code can be revised and
updated over time and it can serve educational use for the team in the
future. In its current state at the end of the 2020-2021 season it
remains unfinished and misses crucial features, however everything is
detailed in the README if you would like to take the look (would be
weird if you were reading commit msgs but not the README lol)