Updating 2020 Branch to Current State
Well, this branch's name is a misnomer now, but I won't change it
for now. This code has been updated for our current status for the
Infinite Recharge Invitational we're hosting, and our code is...
completely unfinished.

Shooter code must be finished and PID Tuned
Intake and BallTrack must be configured properly
Limelight must also be configured