Changes After Day 1
After Day 1 of resuming progress on Infinite Recharge, the intake
and ball track subsystems/commands were reinstated as the build
team worked on them.

In addition to this, reorganized some files around to make more
sense. I also added a bs auto command so when we go to the invy we
still get points for at least moving during the sandstorm period

More changes should come in the next few days