Added Jam Protection for RunBallTrack
Created jam protection for the motors by detecting the current
spike. Once the spike is detected, the motors are ran in reverse
for one second in order to counteract it. This has yet to be
properly tested, and I'll edit the code in the event it fails.

In addition, I began adding some documentation for the code