Structure and BallTrack Updates
A lot of changes occurred today, so read through this list to be
up to date.

The first minor thing was clearing some whitespace and setting
project wide vscode settings so everyone who works on this (in
theory) should have close to identical formatting visible.

Second was the dissolution of the auto unjamming code. We made
edits and got very close to actually solving it, but Mr. Wilson
didn't see any reason to continue sinking time into it given our
current restraints.

Third, I created code to move the balls to the next
position in the shooter. This is still WIP, but most likely will
follow a recursive model.

Finally, I created some ParallelCommandGroups for the intake and
shooting commands. This will let us combine commands in order to
properly execute them.

Sidenote, I'm starting to play around with PathWeaver as well, but
we won't implement it for this robot. We probably will begin
testing it with this robot though.

Long commit over. Miles is the best member of the marching band <3