BallTrack fully functional + Organization
Major update from this commit is that our BallTrack is now fully
functioning. On Friday, we will be able to finish the Intake
system hopefully, as we will be able to test the system with its
required pneumatics.

I organized the code with Command Groups which should be utilized
soon hopefully. Also, added a note about REV Robotics motors.