Changes for Intake
This is first part of two part commit. This is for the intake and
balltrack system. Basically now the system is functioning next to
properly. Most likely won't implement auto unjamming for now, but
it works well enough without it.

Also added the pneumatics for the system, and while they don't
fully work properly, they work well enough for competition. We'll
fix this later on.