Shooting Updates
Messed up the commit, so I'm just making the change for this file, my bad
We have started tuning the shooter PID system, and we are going to finish
it by tomorrow. We have all the other bugs as well to sort out so it'll be fun.

Limelight is now functioning and tuned (it seems), and we'll implement the
necessary methods soon.

We should finish all of our code tomorrow for this revision.